Extracted model training out of PercolatorFdr into PercolatorTrainer
* RunPercolator, per-fold SVM and GBT training, positive-set selection, the cost
  grid search and the Granholm cross-fold calibration
* PercolatorFdr.cs 2,845 -> 1,561 lines and now holds only FDR math; 1,291 moved
  lines verified verbatim
* Six more q-value helpers widened to internal, which measures how far the
  train-score-FDR monolith reaches rather than adding new coupling
* Stellar regression byte-identical on modes 1, 2 and 3

See ai/todos/active/TODO-20260727_osprey_percolatorfdr_decomposition.md
